Revert "refactor : dto 클래스들 directory 분리"
This reverts commit 591d67ccaf8913b108bab204554c797ede2ac18b.
