+++ src/main/java/com/example/newsfeed/dto/CommentResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.newsfeed.dto.commentdto;
+package com.example.newsfeed.dto;
 
 import com.example.newsfeed.entity.Comment;
 import com.example.newsfeed.entity.Menu;
