+++ src/main/java/com/example/newsfeed/dto/PwdCheckRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.newsfeed.dto.userdto;
+package com.example.newsfeed.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
