+++ src/main/java/com/example/newsfeed/dto/SignupRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.newsfeed.dto.userdto;
+package com.example.newsfeed.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
