+++ src/main/java/com/example/newsfeed/dto/UserRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.newsfeed.dto.userdto;
+package com.example.newsfeed.dto;
 
 import lombok.Getter;
